backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
supabase==2.4.3
anthropic==0.26.0
httpx==0.27.0
python-telegram-bot==21.1.1
beautifulsoup4==4.12.3
lxml>=5.2.1
apscheduler==3.10.4
feedparser==6.0.11
slowapi==0.1.9
